+++ CHANGELOG.md
@@ -16,11 +16,11 @@ Changelog for NeoFS Node
 ### Changed
 - FSTree storage now uses more efficient and safe temporary files under Linux (#2566)
 - BoltDB open timeout increased from 100ms to 1s (#2499)
-- Default Policer's handled objects cache size (#2600)
 - Internal container cache size from 10 to 1000 (#2600)
 
 ### Removed
 - deprecated `no-precheck` flag of `neofs-cli container set-eacl` (#2496)
+- Recently-handled objects Policer's cache (#2600)
 
 ### Updated
 - Update minimal supported Go version up to v1.19 (#2485)